fix backend lint as reported in https://drone.gitea.io/go-gitea/gitea/28939/1/4
use log.Error instead of log.Info

fix comments

fix eslint

fix goimport order

add migrations for project priorities

fix _repository.less lint errors

update swagger

fixes according to @6543 suggestions

s/NotInProjectID/ExcludeProjectID/g

update boards priority in a single transaction as per @zeripath


update swagger

Clean up the indenting


use tabs


use tabs

add loadRepository as per @zeripath suggestion

move projects css to features/projects.less
as per @silverwind request
https://github.com/go-gitea/gitea/pull/12506#pullrequestreview-468681403

use transaction to update projects issues priorities
as per @zeripath suggestion
https://github.com/go-gitea/gitea/pull/12506#discussion_r471497361

make import fit on one line

make use of loadRepository

indent html properly
as per suggestion by @silverwind
https://github.com/go-gitea/gitea/pull/12506#discussion_r471665814

remove non-working code as per @zeripath suggestion
https://github.com/go-gitea/gitea/pull/12506#discussion_r471562825

remove non-used code
as per @zeripath suggestion
https://github.com/go-gitea/gitea/pull/12506#discussion_r471517773

fix syntax, remove content-type on request
https://github.com/go-gitea/gitea/pull/12506#discussion_r471763018

use closest instead of parent
as per @silverwind suggestion
https://github.com/go-gitea/gitea/pull/12506#pullrequestreview-468821354

use for instead of forEach because @silverwind

use for instead of forEach because @silverwind

https://github.com/go-gitea/gitea/pull/12506#pullrequestreview-468820176

use // falls through as per @silverwind

https://github.com/go-gitea/gitea/pull/12506#discussion_r471766793

listen to body keyup only on project page

https://github.com/go-gitea/gitea/pull/12506#pullrequestreview-468822919

Update models/project_board.go


move all issue related code to its own util file
and re-use those functions in index.js and projects.js
as per @silverwind and @zeripath

Update models/project_issue.go


fix linting

Update web_src/js/features/issuesutil.js


Update web_src/js/features/issuesutil.js


Update web_src/js/features/issuesutil.js


Update web_src/js/features/projects.js


Update web_src/js/features/projects.js


Update models/project_issue.go

Update models/issue.go


Update web_src/js/features/issuesutil.js


various js fix per @silverwind suggestions

various fixes according to @silverwind and @zeripath

use await for fetch when possible

specify table name as per @lafriks suggestion

specify table name as per @lafriks suggestion in more queries

https://github.com/go-gitea/gitea/pull/12506#pullrequestreview-470332543

https://github.com/go-gitea/gitea/pull/12506#discussion_r472931579

Update templates/repo/projects/view.tmpl


Update templates/repo/projects/view.tmpl


Update web_src/js/features/projects.js


fix project_issue table name in query
make another fetch async as per @silverwind

Update models/project.go


Update models/project.go


more and rename UpdateBoards functions as per @6543

Update models/project.go


Update models/project.go


add / to route

fixup! fix backend lint as reported in https://drone.gitea.io/go-gitea/gitea/28939/1/4

Update models/project.go


Update routers/repo/issue.go


Update web_src/less/features/projects.less


fix uneeded column in query and remove css class
as suggested by @silverwind

Update models/project.go


add X to close sidebar and other css suggestions by @silverwind

add emoji to project issue titles

use issue.* instead as suggested by @zeripath
still waiting for xorm v1.0.5 to update go.mod

fix project_issue query for now

fix card details css
