+++ routers/api/v1/repo/issue.go
@@ -51,7 +51,7 @@ func SearchIssues(ctx *context.APIContext) {
 	//   description: repository to prioritize in the results
 	//   type: integer
 	//   format: int64
-	// - name: not_in_project_id
+	// - name: exclude_project_id
 	//   in: query
 	//   description: project to exclude from search when searching issues to add to a project
 	//   type: integer
@@ -64,6 +64,10 @@ func SearchIssues(ctx *context.APIContext) {
 	//   in: query
 	//   description: page number of requested issues
 	//   type: integer
+	// - name: render_emoji_title
+	//   in: query
+	//   description: render emojis in issue title
+	//   type: string
 	// responses:
 	//   "200":
 	//     "$ref": "#/responses/IssueList"
@@ -76,6 +80,16 @@ func SearchIssues(ctx *context.APIContext) {
 		isClosed = util.OptionalBoolNone
 	default:
 		isClosed = util.OptionalBoolFalse
+	}
+
+	var renderEmojiTitle util.OptionalBool
+	switch ctx.Query("render_emoji_title") {
+	case "true":
+		renderEmojiTitle = util.OptionalBoolTrue
+	case "false":
+		renderEmojiTitle = util.OptionalBoolFalse
+	default:
+		renderEmojiTitle = util.OptionalBoolFalse
 	}
 
 	// find repos user can access (for issue search)
@@ -163,7 +177,8 @@ func SearchIssues(ctx *context.APIContext) {
 			SortType:           "priorityrepo",
 			PriorityRepoID:     ctx.QueryInt64("priority_repo_id"),
 			IsPull:             isPull,
-			NotInProjectID:     ctx.QueryInt64("not_in_project_id"),
+			ExcludeProjectID:   ctx.QueryInt64("exclude_project_id"),
+			RenderEmojiTitle:   renderEmojiTitle,
 		}
 
 		if issues, err = models.Issues(issuesOpt); err != nil {
@@ -234,6 +249,10 @@ func ListIssues(ctx *context.APIContext) {
 	//   in: query
 	//   description: page size of results
 	//   type: integer
+	// - name: render_emoji_title
+	//   in: query
+	//   description: render emojis in issue title
+	//   type: string
 	// responses:
 	//   "200":
 	//     "$ref": "#/responses/IssueList"
@@ -246,6 +265,16 @@ func ListIssues(ctx *context.APIContext) {
 		isClosed = util.OptionalBoolNone
 	default:
 		isClosed = util.OptionalBoolFalse
+	}
+
+	var renderEmojiTitle util.OptionalBool
+	switch ctx.Query("render_emoji_title") {
+	case "true":
+		renderEmojiTitle = util.OptionalBoolTrue
+	case "false":
+		renderEmojiTitle = util.OptionalBoolFalse
+	default:
+		renderEmojiTitle = util.OptionalBoolFalse
 	}
 
 	var issues []*models.Issue
@@ -320,14 +349,15 @@ func ListIssues(ctx *context.APIContext) {
 	// This would otherwise return all issues if no issues were found by the search.
 	if len(keyword) == 0 || len(issueIDs) > 0 || len(labelIDs) > 0 {
 		issuesOpt := &models.IssuesOptions{
-			ListOptions:    listOptions,
-			RepoIDs:        []int64{ctx.Repo.Repository.ID},
-			IsClosed:       isClosed,
-			IssueIDs:       issueIDs,
-			LabelIDs:       labelIDs,
-			MilestoneIDs:   mileIDs,
-			IsPull:         isPull,
-			NotInProjectID: ctx.QueryInt64("not_in_project_id"),
+			ListOptions:      listOptions,
+			RepoIDs:          []int64{ctx.Repo.Repository.ID},
+			IsClosed:         isClosed,
+			IssueIDs:         issueIDs,
+			LabelIDs:         labelIDs,
+			MilestoneIDs:     mileIDs,
+			IsPull:           isPull,
+			ExcludeProjectID: ctx.QueryInt64("exclude_project_id"),
+			RenderEmojiTitle: renderEmojiTitle,
 		}
 
 		if issues, err = models.Issues(issuesOpt); err != nil {
