+++ web_src/less/_repository.less
@@ -3169,82 +3169,6 @@ tbody.commit-list {
   padding-top: 0 !important;
 }
 
-.board {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  overflow-x: auto;
-  margin: 0 .5em;
-}
-
-.board-column {
-  background-color: rgba(0, 0, 0, .05) !important;
-  border: 1px solid var(--color-secondary) !important;
-  margin: 0 .5rem !important;
-  padding: .5rem !important;
-  width: 320px;
-  height: 60vh;
-  overflow-y: scroll;
-  flex: 0 0 auto;
-  overflow: visible;
-  display: flex;
-  flex-direction: column;
-}
-
-.board-column-header {
-  display: flex;
-  justify-content: space-between;
-}
-
-.board-label {
-  background: none !important;
-  line-height: 1.25 !important;
-}
-
-.board-column > .cards {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  margin: 0 !important;
-  padding: 0 !important;
-
-  .card .meta > a.milestone {
-    color: #999999;
-  }
-}
-
-.board-column > .divider {
-  margin: 5px 0;
-}
-
-.board-column:first-child {
-  margin-left: auto !important;
-}
-
-.board-column:last-child {
-  margin-right: auto !important;
-}
-
-.board-card {
-  margin: 3px !important;
-  width: auto !important;
-  background-color: #fff;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.board-card .header {
-  font-size: 1.1em !important;
-}
-
-.board-card .content {
-  padding: 5px 8px !important;
-}
-
-.board-card .extra.content {
-  padding: 5px 8px !important;
-}
-
 td.blob-excerpt {
   background-color: #fafafa;
 }
@@ -3341,43 +3265,4 @@ td.blob-excerpt {
 
 .migrate .cards .card {
   text-align: center;
-}
-
-#current-card-details {
-  visibility: hidden;
-  position: absolute;
-  right: 0px;
-  background-color: #fff;
-  width: 200px;
-  overflow-y: scroll;
-  white-space: normal;
-  white-space: initial;
-  z-index: 100;
-  width: 280px;
-  margin: 0 0.5rem !important;
-  padding: 0.5rem !important;
-  width: 320px;
-  height: 60vh;
-  overflow-y: scroll;
-  flex: 0 0 auto;
-  overflow: scroll;
-  display: flex;
-  flex-direction: column;
-}
-
-#issue-search {
-    margin: 1rem 0.8rem 0 0.1rem;
-    width: 300px;
-    position: absolute;
-    right: 0px;
-    z-index: 1000;
-    padding: 0.5rem !important;
-    flex: 0 0 auto;
-    display: flex;
-    flex-direction: column;
-}
-
-#issue-results {
-    margin-top: 4rem;
-    padding: 0.5rem;
 }
