+++ web_src/less/index.less
@@ -3,6 +3,7 @@
 @import "./features/gitgraph.less";
 @import "./features/animations.less";
 @import "./features/heatmap.less";
+@import "./features/projects.less";
 @import "./markdown/mermaid.less";
 
 @import "_svg";
